Fix signal handlers by moving shutdown logic out of handler.
